+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Sanitize empty symbols in language columns (#1299, @odaniel1).
 * Handle cases where `NROW()` throws an error (#1300, `julian-tagell` on Stack Overflow).
+* Prohibit dynamic branching over non-branching dynamic files (#1302, @djbirke).
 
 # Version 7.12.4